+++ lib/samly/auth_handler.ex
@@ -3,9 +3,10 @@ defmodule Samly.AuthHandler do
 
   require Logger
   import Plug.Conn
-  alias Samly.{Assertion, IdpData, Helper, State, Subject}
-
   import Samly.RouterUtil, only: [ensure_sp_uris_set: 2, send_saml_request: 5, redirect: 3]
+
+  alias Plug.Conn
+  alias Samly.{Assertion, IdpData, Helper, State, Subject}
 
   @sso_init_resp_template """
   <body onload=\"document.forms[0].submit()\">
@@ -25,6 +26,12 @@ defmodule Samly.AuthHandler do
   </body>
   """
 
+  def initiate_sso_req(%Conn{params: %Conn.Unfetched{}} = conn) do
+    conn
+    |> fetch_query_params()
+    |> initiate_sso_req()
+  end
+
   def initiate_sso_req(conn) do
     import Plug.CSRFProtection, only: [get_csrf_token: 0]
 
@@ -41,21 +48,21 @@ defmodule Samly.AuthHandler do
     ]
 
     conn
-    |> put_resp_header("Content-Type", "text/html")
+    |> put_resp_header("content-type", "text/html")
     |> send_resp(200, EEx.eval_string(@sso_init_resp_template, opts))
   end
 
-  def send_signin_req(conn) do
+  def send_signin_req(conn, state \\ Application.get_env(:samly, :state_provider, State.Ets)) do
     %IdpData{id: idp_id} = idp = conn.private[:samly_idp]
     %IdpData{esaml_idp_rec: idp_rec, esaml_sp_rec: sp_rec} = idp
     sp = ensure_sp_uris_set(sp_rec, conn)
 
     target_url = (conn.params["target_url"] || "/") |> URI.decode_www_form()
-    nameid = get_session(conn, "samly_nameid")
 
-    case State.get_by_nameid(nameid) do
-      {^nameid, %Assertion{idp_id: ^idp_id}} ->
-        conn |> redirect(302, target_url)
+    case state.get(conn, "samly_assertion") do
+      {_ ,%Assertion{idp_id: ^idp_id}}->
+        conn
+        |> redirect(302, target_url)
 
       _ ->
         relay_state = State.gen_id()
@@ -73,37 +80,30 @@ defmodule Samly.AuthHandler do
           relay_state |> URI.encode_www_form()
         )
     end
-
-    # rescue
-    #   error ->
-    #     Logger.error("#{inspect error}")
-    #     conn |> send_resp(500, "request_failed")
   end
 
-  def send_signout_req(conn) do
+  def send_signout_req(conn, state \\ Application.get_env(:samly, :state_provider, State.Ets)) do
     %IdpData{id: idp_id} = idp = conn.private[:samly_idp]
     %IdpData{esaml_idp_rec: idp_rec, esaml_sp_rec: sp_rec} = idp
     sp = ensure_sp_uris_set(sp_rec, conn)
 
     target_url = (conn.params["target_url"] || "/") |> URI.decode_www_form()
-    nameid = get_session(conn, "samly_nameid")
 
-    case State.get_by_nameid(nameid) do
-      {^nameid, %Assertion{idp_id: ^idp_id, authn: authn, subject: subject}} ->
+    case state.get(conn,"samly_assertion") do
+      {_, %Assertion{idp_id: ^idp_id, authn: authn, subject: subject}} ->
         session_index = Map.get(authn, "session_index", "")
         subject_rec = Subject.to_rec(subject)
 
         {idp_signout_url, req_xml_frag} =
           Helper.gen_idp_signout_req(sp, idp_rec, subject_rec, session_index)
 
-        State.delete(nameid)
         relay_state = State.gen_id()
 
         conn
+        |> state.delete("samly_assertion")
         |> put_session("target_url", target_url)
         |> put_session("relay_state", relay_state)
         |> put_session("idp_id", idp_id)
-        |> delete_session("samly_nameid")
         |> send_saml_request(
           idp_signout_url,
           idp.use_redirect_for_req,
@@ -114,10 +114,5 @@ defmodule Samly.AuthHandler do
       _ ->
         conn |> send_resp(403, "access_denied")
     end
-
-    # rescue
-    #   error ->
-    #     Logger.error("#{inspect error}")
-    #     conn |> send_resp(500, "request_failed")
   end
 end